fix(auth): OAuth signup guard 경계 정렬
