+++ back/src/main/kotlin/com/back/boundedContexts/member/subContexts/oauthSignup/adapter/persistence/PendingOAuthSignupRepository.kt
@@ -1,6 +1,6 @@
 package com.back.boundedContexts.member.subContexts.oauthSignup.adapter.persistence
 
-import com.back.boundedContexts.member.subContexts.oauthSignup.domain.PendingOAuthSignup
+import com.back.boundedContexts.member.subContexts.oauthSignup.model.PendingOAuthSignup
 import org.springframework.data.jpa.repository.JpaRepository
 
 interface PendingOAuthSignupRepository : JpaRepository<PendingOAuthSignup, Long> {
