+++ back/src/main/kotlin/com/back/boundedContexts/member/subContexts/oauthSignup/adapter/persistence/PendingOAuthSignupRepositoryAdapter.kt
@@ -1,7 +1,7 @@
 package com.back.boundedContexts.member.subContexts.oauthSignup.adapter.persistence
 
 import com.back.boundedContexts.member.subContexts.oauthSignup.application.port.output.PendingOAuthSignupRepositoryPort
-import com.back.boundedContexts.member.subContexts.oauthSignup.domain.PendingOAuthSignup
+import com.back.boundedContexts.member.subContexts.oauthSignup.model.PendingOAuthSignup
 import org.springframework.stereotype.Component
 
 @Component
