+++ back/src/main/kotlin/com/back/boundedContexts/member/subContexts/oauthSignup/application/port/output/PendingOAuthSignupRepositoryPort.kt
@@ -1,6 +1,6 @@
 package com.back.boundedContexts.member.subContexts.oauthSignup.application.port.output
 
-import com.back.boundedContexts.member.subContexts.oauthSignup.domain.PendingOAuthSignup
+import com.back.boundedContexts.member.subContexts.oauthSignup.model.PendingOAuthSignup
 
 interface PendingOAuthSignupRepositoryPort {
     fun save(pendingOAuthSignup: PendingOAuthSignup): PendingOAuthSignup
