+++ back/src/main/kotlin/com/back/boundedContexts/member/subContexts/oauthSignup/application/service/OAuthSignupResults.kt
@@ -1,0 +1,16 @@
+package com.back.boundedContexts.member.subContexts.oauthSignup.application.service
+
+import java.time.Instant
+
+data class OAuthSignupPendingStartResult(
+    val provider: String,
+    val pendingToken: String,
+    val expiresAt: Instant,
+)
+
+data class OAuthSignupPendingDetails(
+    val provider: String,
+    val nickname: String,
+    val profileImgUrl: String?,
+    val expiresAt: Instant,
+)
