--- back/src/main/kotlin/com/back/boundedContexts/member/subContexts/oauthSignup/domain/PersistenceModelAliases.kt
@@ -1,3 +1,0 @@
-package com.back.boundedContexts.member.subContexts.oauthSignup.domain
-
-typealias PendingOAuthSignup = com.back.boundedContexts.member.subContexts.oauthSignup.model.PendingOAuthSignup
